# Clear the JSONAllocator after each batch processed